backend/requirements.txt
django-cors-headers==3.13.0
pillow==9.5.0